+++ common/web/types/src/util/compiler-interfaces.ts
@@ -99,8 +99,9 @@ export enum CompilerErrorNamespace {
    */
   Analyzer = 0x6000,
   /**
-   * kmc-kmw errors between 0x7000…0x7FFF; note that some errors generated by
-   * kmc-kmw are from kmc-kmn namespace for legacy reasons
+   * kmc-kmn/kmw-compiler errors between 0x7000…0x7FFF; note that some errors
+   * generated by kmc-kmn/kmw-compiler are from kmc-kmn namespace for legacy
+   * reasons
    */
   KmwCompiler = 0x7000,
 };
