+++ developer/src/kmc-kmn/.eslintrc.cjs
@@ -8,6 +8,7 @@ module.exports = {
       extends: ["../../../common/web/eslint/eslintNoNodeImports.js"],
     }
   ],
+  ignorePatterns: ["test/fixtures/*"],
   rules: {
     "prefer-const": "off", // TODO: enable this once infrastructure is in place and cleanup the problem cases
   },
