+++ developer/src/kmc-kmn/package.json
@@ -62,6 +62,7 @@
     "exclude-after-remap": true,
     "exclude": [
       "src/import/",
+      "src/kmw-compiler",
       "test/"
     ]
   },
