+++ developer/src/kmc-kmn/src/kmw-compiler/compiler-globals.ts
