+++ developer/src/kmc-kmn/src/main.ts
@@ -1,3 +1,4 @@
 
-export { KmnCompilerMessages, CompilerMessages } from './compiler/messages.js';
 export { KmnCompiler } from './compiler/compiler.js';
+export { KmnCompilerMessages, CompilerMessages } from './compiler/messages.js';
+export { KmwCompilerMessages } from './kmw-compiler/messages.js';
