+++ developer/src/kmc-kmn/test/fixtures/kmw/khmer_angkor.kmn
