+++ developer/src/kmc-kmn/tsconfig.json
@@ -8,7 +8,6 @@
     "allowJs": true,
     "paths": {
       "@keymanapp/common-types": ["../../../common/web/types/src/main"],
-      "@keymanapp/kmc-kmw": ["../kmc-kmw/src/main"],
     },
 
   },
@@ -19,6 +18,5 @@
   "references": [
     { "path": "../../../common/web/keyman-version/tsconfig.esm.json" },
     { "path": "../../../common/web/types/" },
-    { "path": "../kmc-kmw/" },
   ]
 }
