--- developer/src/kmc-kmw/README.md
@@ -1,1 +1,0 @@
-# kmc-kmw
