--- developer/src/kmc-kmw/package.json
@@ -1,56 +1,0 @@
-{
-  "name": "@keymanapp/kmc-kmw",
-  "description": "Keyman Developer KMW keyboard compiler",
-  "keywords": [
-    "keyboard",
-    "keyman",
-    "ldml",
-    "unicode"
-  ],
-  "type": "module",
-  "exports": {
-    ".": "./build/src/main.js"
-  },
-  "scripts": {
-    "build": "gosh ./build.sh build",
-    "test": "gosh ./build.sh test",
-    "prepublishOnly": "npm run build"
-  },
-  "author": "Marc Durdin <marc@keyman.com> (https://github.com/mcdurdin)",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/keymanapp/keyman/issues"
-  },
-  "dependencies": {
-    "@keymanapp/common-types": "*",
-    "@keymanapp/keyman-version": "*",
-    "ajv": "^8.11.0",
-    "restructure": "git+https://github.com/keymanapp/dependency-restructure.git#7a188a1e26f8f36a175d95b67ffece8702363dfc",
-    "semver": "^7.3.7",
-    "xml2js": "git+https://github.com/keymanapp/dependency-node-xml2js#535fe732dc408d697e0f847c944cc45f0baf0829"
-  },
-  "devDependencies": {
-    "@types/chai": "^4.1.7",
-    "@types/mocha": "^5.2.7",
-    "@types/node": "^10.14.6",
-    "@types/semver": "^7.3.12",
-    "@types/xml2js": "^0.4.5",
-    "c8": "^7.12.0",
-    "chai": "^4.3.4",
-    "chalk": "^2.4.2",
-    "mocha": "^8.4.0",
-    "ts-node": "^9.1.1",
-    "typescript": "^4.9.5",
-    "@keymanapp/resources-gosh": "*"
-  },
-  "mocha": {
-    "spec": "build/test/**/test-*.js",
-    "require": [
-      "source-map-support/register"
-    ]
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/keymanapp/keyman.git"
-  }
-}
