--- developer/src/kmc-kmw/src/main.ts
@@ -1,2 +1,0 @@
-
-export { WriteCompiledKeyboard } from './compiler/write-compiled-keyboard.js';
