--- developer/src/kmc-kmw/test/README.md
@@ -1,7 +1,0 @@
-Keyman kmn kmw Keyboard Compiler Tests
-======================================
-
-Test
-----
-
-    ../build.sh test
