--- developer/src/kmc-kmw/test/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "extends": "../../kmc/tsconfig.kmc-base.json",
-
-  "compilerOptions": {
-      "rootDir": ".",
-      "rootDirs": ["./", "../src/"],
-      "outDir": "../build/test",
-      "baseUrl": ".",
-      "allowSyntheticDefaultImports": true, // for chai
-      "paths": {
-        "@keymanapp/common-types": ["../../../../common/web/types/src/main"],
-        "@keymanapp/developer-test-helpers": ["../../common/web/test-helpers/index"],
-        "@keymanapp/kmc-kmn": ["../../kmc-kmn/src/main"],
-      },
-    },
-  "include": [
-      "**/test-*.ts",
-      "util.ts"
-  ],
-  "references": [
-      { "path": "../../../../common/web/types/" },
-      { "path": "../../kmc-kmn/" },
-      { "path": "../../common/web/test-helpers/" },
-      { "path": "../" }
-    ]
-}
