--- developer/src/kmc-kmw/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "extends": "../../../tsconfig.esm-base.json",
-
-  "compilerOptions": {
-    "outDir": "build/src/",
-    "rootDir": "src/",
-    "baseUrl": ".",
-    "paths": {
-      "@keymanapp/common-types": ["../../../common/web/types/src/main"],
-    },
-
-  },
-  "include": [
-    "src/**/*.ts"
-  ],
-  "references": [
-    { "path": "../../../common/web/types/" },
-    { "path": "../../../core/include/ldml/"},
-  ]
-}
