+++ developer/src/kmc/build.sh
@@ -22,7 +22,6 @@ builder_describe "Build Keyman Keyboard Compiler kmc" \
   "@/common/web/types" \
   "@/developer/src/kmc-analyze" \
   "@/developer/src/kmc-kmn" \
-  "@/developer/src/kmc-kmw" \
   "@/developer/src/kmc-ldml" \
   "@/developer/src/kmc-model" \
   "@/developer/src/kmc-model-info" \
@@ -116,7 +115,6 @@ readonly PACKAGES=(
   core/include/ldml
   developer/src/kmc-analyze
   developer/src/kmc-kmn
-  developer/src/kmc-kmw
   developer/src/kmc-ldml
   developer/src/kmc-model
   developer/src/kmc-model-info
