+++ developer/src/kmc/package.json
@@ -41,7 +41,6 @@
     "@keymanapp/common-types": "*",
     "@keymanapp/kmc-analyze": "*",
     "@keymanapp/kmc-kmn": "*",
-    "@keymanapp/kmc-kmw": "*",
     "@keymanapp/kmc-ldml": "*",
     "@keymanapp/kmc-model": "*",
     "@keymanapp/kmc-model-info": "*",
