+++ developer/src/kmc/tsconfig.json
@@ -9,7 +9,6 @@
       "@keymanapp/common-types": [ "../../../common/web/types/src/main" ],
       "@keymanapp/kmc-analyze": [ "../kmc-analyze/src/index" ],
       "@keymanapp/kmc-kmn": [ "../kmc-kmn/src/main" ],
-      "@keymanapp/kmc-kmw": [ "../kmc-kmw/src/main" ],
       "@keymanapp/kmc-ldml": [ "../kmc-ldml/src/main" ],
       "@keymanapp/kmc-model": [ "../kmc-model/src/main" ],
       "@keymanapp/kmc-model-info": [ "../kmc-model-info/src/model-info-compiler" ],
@@ -24,7 +23,6 @@
     { "path": "../../../common/web/types" },
     { "path": "../kmc-analyze" },
     { "path": "../kmc-kmn" },
-    { "path": "../kmc-kmw" },
     { "path": "../kmc-ldml" },
     { "path": "../kmc-model" },
     { "path": "../kmc-model-info" },
