+++ tsconfig.esm.json
@@ -16,8 +16,6 @@
     { "path": "./developer/src/kmc-analyze/tsconfig.json" },
     { "path": "./developer/src/kmc-kmn/test/tsconfig.json" },
     { "path": "./developer/src/kmc-kmn/tsconfig.json" },
-    { "path": "./developer/src/kmc-kmw/test/tsconfig.json" },
-    { "path": "./developer/src/kmc-kmw/tsconfig.json" },
     { "path": "./developer/src/kmc-ldml/test/tsconfig.json" },
     { "path": "./developer/src/kmc-ldml/tsconfig.json" },
     { "path": "./developer/src/kmc-model/test/tsconfig.json" },
